{WIP} Use the plugin metadata so we can get plugin details without loading the plugin
The metadata is in JSON format (see qDummyPlugin info.json). In order to also access this info from the plugin interface, we store the JSON as a Qt resource as well and access it in the new intermediate ccDefaultPluginInterface class. (Yes this means we are storing this data twice, but there's no easy way to get at the metadata from within the plugin interface.)

This class reads the JSON internally and satisfies the interface for the data methods getName(), getDescription(), getIcon(), getReferences(), getAuthors(), and getMaintainers(). This means that classes deriving from ccDefaultPluginInterface which use the JSON file no longer need to implement these methods (see qDummyPlugin example).

This is implemented in such a way that all plugins will continue to compile and work, but only those providing the info.json will be visible using QPluginLoader::metaData(). So they don't all have to be changed immediately.